+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.5.2
+* All the classes are generated parsing the official Highcharts API
+* @JS() directive added on top of libraries to make it work with sdk 1.17
+
 ## 0.0.1
 
 * Initial version, created by Stagehand.